refactor: 컨벤션에 맞게 네이밍, 개행 수정 (#18)
