+++ src/main/java/com/team/buddyya/certification/dto/response/CertificationResponse.java
@@ -1,6 +1,4 @@
 package com.team.buddyya.certification.dto.response;
 
-public record CertificationResponse(
-        boolean success
-) {
+public record CertificationResponse(boolean success) {
 }
