+++ src/main/java/com/team/buddyya/common/service/S3UploadService.java
@@ -1,4 +1,4 @@
-package com.team.buddyya.common.infrastructure;
+package com.team.buddyya.common.service;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -7,14 +7,15 @@ import com.team.buddyya.common.exception.CommonExceptionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.UUID;
 
 @RequiredArgsConstructor
-@Component
-public class S3UploadManager {
+@Service
+public class S3UploadService {
 
     private final AmazonS3Client amazonS3;
